Fixed bug

+++ backend/UCN-Semester-3-project/RecipesData/Database/SwipedRecipeDatabaseAccess.cs
@@ -156,12 +156,12 @@ namespace RecipesData.Database
         /// </summary>
         /// <param name="id">The Guid of the Recipe that is associated with this SwipedRecipe</param>
         /// <returns>True if the removal was successful</returns>
-        public bool DeleteSR(Guid id)
+        public bool DeleteSR(Guid id, Guid userId)
         {
             bool removed = false;
             string guidString = id.ToString();
 
-            string query = "delete from SwipedRecipe where recipeId = @id";
+            string query = "delete from SwipedRecipe where recipeId = @id and userId = @userId";
 
             using (SqlConnection connection = new SqlConnection(_connectionString))
             {
@@ -170,6 +170,7 @@ namespace RecipesData.Database
                 {
                     command.CommandText = query;
                     command.Parameters.AddWithValue("@id", id);
+                    command.Parameters.AddWithValue("@userId", userId);
 
                     command.ExecuteNonQuery();
                     removed = true;
